Module Ast_derive_abstract

type abstractKind =

| Not_abstract

| Light_abstract

| Complex_abstract

;

let isAbstract: list(Ast_payload.action) => abstractKind;

if only abstract happens true if abstract does not appear false if abstract happens with other, raise exception

let handleTdclsInStr: light:bool => Asttypes.rec_flag => list(Melange_compiler_libs.Parsetree.type_declaration) => Melange_compiler_libs.Parsetree.structure;

let handleTdclsInSig: light:bool => Asttypes.rec_flag => list(Melange_compiler_libs.Parsetree.type_declaration) => Melange_compiler_libs.Parsetree.signature;