Module Ast_payload

A utility module used when destructuring parsetree attributes, used for compiling FFI attributes and built-in ppx

type t = Melange_compiler_libs.Parsetree.payload;

type lid = Asttypes.loc(string);

type label_expr = (lid, Melange_compiler_libs.Parsetree.expression);

type action = (lid, option(Melange_compiler_libs.Parsetree.expression));

let is_single_string: t => option((string, option(string)));

let is_single_string_as_ast: t => option(Melange_compiler_libs.Parsetree.expression);

let is_single_int: t => option(int);

let raw_as_string_exp_exn: kind:Js_raw_info.raw_kind => ?is_function:Stdlib.ref(bool) => t => option(Melange_compiler_libs.Parsetree.expression);

Convert %raw into expression

let as_core_type: Melange_compiler_libs.Location.t => t => Melange_compiler_libs.Parsetree.core_type;

let as_ident: t => option(Asttypes.loc(Longident.t));

let assert_strings: Melange_compiler_libs.Location.t => t => list(string);

as a record or empty it will accept

[@@@bs.config ]

or

[@@@bs.config no_export ] 

or

[@@@bs.config { property  .. } ]

Note that we only

{ flat_property}

below is not allowed

{M.flat_property}

let ident_or_record_as_config: Melange_compiler_libs.Location.t => t => list(action);

let assert_bool_lit: Melange_compiler_libs.Parsetree.expression => bool;

let empty: t;

let table_dispatch: Map_string.t((option(Melange_compiler_libs.Parsetree.expression) => 'a)) => action => option('a);

let unrecognizedConfigRecord: Melange_compiler_libs.Location.t => string => unit;

Report to the user, as a warning, that the bs-attribute parser is bailing out. (This is to allow external ppx, like ppx_deriving, to pick up where the builtin ppx leave off.)