Module Belt.HashMap

Belt.HashMap

The top level provides generic mutable hash map operations.

It also has two specialized inner modules Belt.HashMap.Int and Belt.HashMap.String

A mutable Hash map which allows customized Hash behavior.

All data are parameterized by not its only type but also a unique identity in the time of initialization, so that two HashMaps of ints initialized with different hash functions will have different type.

For example:

type t = int
module I0 =
  (val Belt.Id.hashableU
      ~hash:(fun[\@bs] (a : t)  -> a & 0xff_ff)
      ~eq:(fun[\@bs] a b -> a = b)
  )
let s0 : (_, string,_) t = make ~hintSize:40 ~id:(module I0)
module I1 =
  (val Belt.Id.hashableU
      ~hash:(fun[\@bs] (a : t)  -> a & 0xff)
      ~eq:(fun[\@bs] a b -> a = b)
  )
let s1 : (_, string,_) t  = make ~hintSize:40 ~id:(module I1)

The invariant must be held: for two elements who are equal, their hashed value should be the same

Here the compiler would infer s0 and s1 having different type so that it would not mix.

val s0 :  (int, I0.identity) t
val s1 :  (int, I1.identity) t

We can add elements to the collection:

let () =
  add s1 0 "3";
  add s1 1 "3"

Since this is an mutable data strucure, s1 will contain two pairs.

module Int: { ... };

Specalized when key type is int, more efficient than the generic type

module String: { ... };

Specalized when key type is string, more efficient than the generic type

type t('key, 'value, 'id);

The type of hash tables from type 'key to type 'value.

type id('a, 'id) = (module Belt__.Belt_Id.Hashable with type identity = 'id and type t = 'a);

let make: hintSize:int => id:id('key, 'id) => t('key, 'value, 'id);

let clear: t('key, 'value, 'id) => unit;

Empty a hash table.

let isEmpty: t(_, _, _) => bool;

let set: t('key, 'value, 'id) => 'key => 'value => unit;

set tbl k v if k does not exist, add the binding k,v, otherwise, update the old value with the new v

let copy: t('key, 'value, 'id) => t('key, 'value, 'id);

let get: t('key, 'value, 'id) => 'key => option('value);

let has: t('key, 'value, 'id) => 'key => bool;

has tbl x checks if x is bound in tbl.

let remove: t('key, 'value, 'id) => 'key => unit;

let forEachU: t('key, 'value, 'id) => Js.Fn.arity2(('key => 'value => unit)) => unit;

let forEach: t('key, 'value, 'id) => ('key => 'value => unit) => unit;

forEach tbl f applies f to all bindings in table tbl. f receives the key as first argument, and the associated value as second argument. Each binding is presented exactly once to f.

let reduceU: t('key, 'value, 'id) => 'c => Js.Fn.arity3(('c => 'key => 'value => 'c)) => 'c;

let reduce: t('key, 'value, 'id) => 'c => ('c => 'key => 'value => 'c) => 'c;

reduce tbl init f computes (f kN dN ... (f k1 d1 init)...), where k1 ... kN are the keys of all bindings in tbl, and d1 ... dN are the associated values. Each binding is presented exactly once to f.

The order in which the bindings are passed to f is unspecified. However, if the table contains several bindings for the same key, they are passed to f in reverse order of introduction, that is, the most recent binding is passed first.

let keepMapInPlaceU: t('key, 'value, 'id) => Js.Fn.arity2(('key => 'value => option('value))) => unit;

let keepMapInPlace: t('key, 'value, 'id) => ('key => 'value => option('value)) => unit;

let size: t(_, _, _) => int;

size tbl returns the number of bindings in tbl. It takes constant time.

let toArray: t('key, 'value, 'id) => array(('key, 'value));

let keysToArray: t('key, _, _) => array('key);

let valuesToArray: t(_, 'value, _) => array('value);

let fromArray: array(('key, 'value)) => id:id('key, 'id) => t('key, 'value, 'id);

let mergeMany: t('key, 'value, 'id) => array(('key, 'value)) => unit;

let getBucketHistogram: t(_, _, _) => array(int);

let logStats: t(_, _, _) => unit;