Module Ext_list

let map_combine: list('a) => list('b) => ('a => 'c) => list(('c, 'b));

let map_combine_array: array('a) => list('b) => ('a => 'c) => list(('c, 'b));

let map_combine_array_append: array('a) => list('b) => list(('c, 'b)) => ('a => 'c) => list(('c, 'b));

let map_snd: list(('a, 'b)) => ('b => 'c) => list(('a, 'c));

let map_last: list('a) => (bool => 'a => 'b) => list('b);

map_last f xs will pass true to f for the last element, false otherwise. For empty list, it returns empty

let fold_left_with_offset: list('a) => 'acc => int => ('a => 'acc => int => 'acc) => 'acc;

let same_length: list('a) => list('b) => bool;

let split_at: list('a) => int => (list('a), list('a));

split_at n l will split l into two lists a,b, a will be of length n, otherwise, it will raise

let split_at_last: list('a) => (list('a), 'a);

split_at_last l is equivalent to split_at (List.length l - 1) l

let length_ge: list('a) => int => bool;

length xs = length ys + n 

input n should be positive TODO: input checking

let length_larger_than_n: list('a) => list('a) => int => bool;

let stable_group: list('a) => ('a => 'a => bool) => list(list('a));

stable_group eq lst Example: Input:

stable_group (=) [1;2;3;4;3]

Output:

[[1];[2];[4];[3;3]]

TODO: this is O(n^2) behavior which could be improved later

let rev_iter: list('a) => ('a => unit) => unit;

let for_all2_no_exn: list('a) => list('b) => ('a => 'b => bool) => bool;

for_all2_no_exn p xs ys returns true if all predicates satisfied, false otherwise or when the list lengths are not equal.

let split_map: list('a) => ('a => ('b, 'c)) => (list('b), list('c));

f is applied follow the list order