Module Ext_pp_scope

Scope type to improve identifier name printing Defines scope type t, so that the pretty printer would print more beautiful code: print identifer instead of identifier$1234 when it can

type t

val empty : t

val print : Stdlib.Format.formatter -> t -> unit

val sub_scope : t -> Set_ident.t -> t

val merge : t -> Set_ident.t -> t

val str_of_ident : t -> Ident.t -> string * t

val ident : t -> Ext_pp.t -> Ident.t -> t