Module Js.Exn

Provide utilities for dealing with Js exceptions

type t;

type exn +=

| Error(t)

;

let asJsExn: exn => option(t);

let stack: t => option(string);

let message: t => option(string);

let name: t => option(string);

let fileName: t => option(string);

let isCamlExceptionOrOpenVariant: 'a => bool;

internal use only

let anyToExnInternal: 'a => exn;

* anyToExnInternal obj will take any value obj and wrap it * in a Js.Exn.Error if given value is not an exn already. If * obj is an exn, it will return obj without any changes. * * This function is mostly useful for cases where you want to unify a type of a value * that potentially is either exn, a JS error, or any other JS value really (e.g. for * a value passed to a Promise.catch callback) * * IMPORTANT: This is an internal API and may be changed / removed any time in the future. * *

*   switch (Js.Exn.unsafeAnyToExn("test")) {
*   | Js.Exn.Error(v) =>
*     switch (Js.Exn.message(v)) {
*     | Some(str) => Js.log("We won't end up here")
*     | None => Js.log2("We will land here: ", v)
*     }
*   }
* 

* *

let raiseError: string => 'a;

Raise Js exception Error object with stacktrace

let raiseEvalError: string => 'a;

let raiseRangeError: string => 'a;

let raiseReferenceError: string => 'a;

let raiseSyntaxError: string => 'a;

let raiseTypeError: string => 'a;

let raiseUriError: string => 'a;