Module Env.Unscoped

alert dangerous It is unsafe to use the common [Path.same] function in contexts where [Unscoped.with_pairs] has been used to enrich the environment with unscoped equalities; [Path.equiv] must be used instead. See [ctype.ml] for an example of careful usage of [Unscoped], by locally shadowing [Path] to hide [Path.same].

let with_pairs: list((Ident.Unscoped.t, Ident.Unscoped.t)) => t => t;

let get_pairs: t => list((Ident.Unscoped.t, Ident.Unscoped.t));

let path_equiv: t => Path.t => Path.t => bool;