Module Load_path.Dir

type t;

Represent one directory in the load path.

let create: string => t;

let path: t => string;

let files: t => list(string);

All the files in that directory. This doesn't include files in sub-directories of this directory.

let find: t => string => option(string);

find dir fn returns the full path to fn in dir.

let find_uncap: t => string => option(string);

As find, but search also for uncapitalized name, i.e. if name is Foo.ml, either /path/Foo.ml or /path/foo.ml may be returned.