Module Stdlib.Array

Extensions to the Array module

let exists2: ('a => 'b => bool) => array('a) => array('b) => bool;

Same as Array.exists2 from the standard library.

let for_alli: (int => 'a => bool) => array('a) => bool;

Same as Array.for_all from the standard library, but the function is applied with the index of the element as first argument, and the element itself as second argument.

let all_somes: array(option('a)) => option(array('a));