Module Melange_compiler_libs.Printtyp

Printing functions

type namespace := Shape.Sig_component_kind.t;

let namespaced_ident: namespace => Ident.t => string;

let string_of_path: Path.t => string;

let strings_of_paths: namespace => list(Path.t) => list(string);

Print a list of paths, using the same naming context to avoid name collisions

let printed_signature: string => Stdlib.Format.formatter => Types.signature => unit;

printed_signature sourcefile ppf sg print the signature sg of sourcefile with potential warnings for name collisions

module type Printers := { ... };

module Doc: Printers with type printer('a) := Format_doc.printer('a);

For compatibility with Format printers

include Printers with type printer('a) := Format_doc.format_printer('a);

let wrap_printing_env: error:bool => Env.t => (unit => 'a) => 'a;

Call the function using the environment for type path shortening This affects all the printing functions below Also, if ~error:true, then disable the loading of cmis

let longident: Format_doc.format_printer(Longident.t);

let ident: Format_doc.format_printer(Ident.t);

let path: Format_doc.format_printer(Path.t);

let type_path: Format_doc.format_printer(Path.t);

Print a type path taking account of -short-paths. Calls should be within wrap_printing_env.

let type_expr: Format_doc.format_printer(Types.type_expr);

Print out a type. This will pick names for type variables, and will not reuse names for common type variables shared across multiple type expressions. (It will also reset the printing state, which matters for other type formatters such as prepared_type_expr.) If you want multiple types to use common names for type variables, see Out_type.prepare_for_printing and Out_type.prepared_type_expr.

let type_scheme: Format_doc.format_printer(Types.type_expr);

let shared_type_scheme: Format_doc.format_printer(Types.type_expr);

shared_type_scheme is very similar to type_scheme, but does not reset the printing context first. This is intended to be used in cases where the printing should have a particularly wide context, such as documentation generators; most use cases, such as error messages, have narrower contexts for which type_scheme is better suited.

let type_expansion: Out_type.type_or_scheme => Format_doc.format_printer(Errortrace.expanded_type);

let label: Format_doc.format_printer(Types.label_declaration);

let constructor: Format_doc.format_printer(Types.constructor_declaration);

let constructor_arguments: Format_doc.format_printer(Types.constructor_arguments);

let extension_constructor: Ident.t => Format_doc.format_printer(Types.extension_constructor);

Prints extension constructor with the type signature: type ('a, 'b) bar += A of float

let extension_only_constructor: Ident.t => Format_doc.format_printer(Types.extension_constructor);

Prints only extension constructor without type signature: A of float

let value_description: Ident.t => Format_doc.format_printer(Types.value_description);

let type_declaration: Ident.t => Format_doc.format_printer(Types.type_declaration);

let modtype_declaration: Ident.t => Format_doc.format_printer(Types.modtype_declaration);

let class_declaration: Ident.t => Format_doc.format_printer(Types.class_declaration);

let cltype_declaration: Ident.t => Format_doc.format_printer(Types.class_type_declaration);

let modtype: Format_doc.format_printer(Types.module_type);

let signature: Format_doc.format_printer(Types.signature);

let class_type: Format_doc.format_printer(Types.class_type);