Module Subst.Unsafe

type t = subst(unsafe);

Unsafe substitutions introduced by with constraints, local substitutions (type t := int * int) or recursive module check.

let add_modtype: Ident.t => Types.module_type => subst('any) => t;

Replacing a module type name S by a non-path signature is unsafe as the packed module type (module S) becomes ill-formed.

let add_modtype_path: Path.t => Types.module_type => subst('any) => t;

let add_type_path: Path.t => Path.t => t => t;

Deep editing inside a module type require to retypecheck the module, for applicative functors in path and module aliases.

let add_type_function: Path.t => params:list(Types.type_expr) => body:Types.type_expr => t => t;

let add_module_path: Path.t => Path.t => t => t;

type error =

| Fcm_type_substituted_away(Path.t, Types.module_type)

;

type res('a) := Stdlib.result('a, error);

let type_declaration: t => Types.type_declaration => res(Types.type_declaration);

let signature_item: scoping => t => Types.signature_item => res(Types.signature_item);

let signature: scoping => t => Types.signature => res(Types.signature);

let compose: t => t => res(t);

Composition of substitutions is eager and fails when the two substitution are incompatible, for example module type t := sig end is not compatible with module type s := sig type t=(module t) end