Module Melstd

module Array: { ... };

module Filename: { ... };

module Hash: { ... };

module Hash_gen: { ... };

module Hash_int: { ... };

module Hash_set: { ... };

Ideas are based on Hash, however, Hash.add does not really optimize and has a bad semantics for Hash_set, This module fixes the semantics of add. remove is not optimized since it is not used too much

module Hash_set_gen: { ... };

module Hash_set_ident_mask: { ... };

module Ident: { ... };

A wrapper around Ident module in compiler-libs

module Int_vec_vec: { ... };

module Js_suffix: { ... };

module List: { ... };

module Loc: { ... };

module Log: { ... };

A Poor man's logging utility

module Map_gen: { ... };

module Map_int: { ... };

module Modulename: { ... };

module Module_system: { ... };

module Ordered_hash_map_local_ident: { ... };

module Polyvariant: { ... };

module Path: { ... };

module Js_pp: { ... };

A simple pretty printer

module Scc: { ... };

module String: { ... };

Extension to the standard library String module, fixed some bugs like avoiding locale sensitivity

module Vec: { ... };

module Vec_int: { ... };

module Set_int: { ... };

module Set_gen: { ... };