Module Melstd.Hash_set_ident_mask

type ident = Ident.t;

Based on hash_set specialized for mask operations

type t;

let create: int => t;

let add_unmask: t => ident => unit;

let mask_and_check_all_hit: t => ident => bool;

check_mask h key if key exists mask it otherwise nothing return true if all keys are masked otherwise false

let iter_and_unmask: t => (ident => bool => unit) => unit;

iter_and_unmask f h iterating the collection and mask all idents, dont consul the collection in function f TODO: what happens if an exception raised in the callback, would the hashtbl still be in consistent state?