Module Melstd.List

include module type of struct include Stdlib.StdLabels.List end;

type t(!'a) = list('a) =

| []

| ::('a, list('a))

;

let length: list('a) => int;

let compare_lengths: list('a) => list('b) => int;

let compare_length_with: list('a) => len:int => int;

let is_empty: list('a) => bool;

let cons: 'a => list('a) => list('a);

let hd: list('a) => 'a;

let tl: list('a) => list('a);

let nth: list('a) => int => 'a;

let nth_opt: list('a) => int => option('a);

let rev: list('a) => list('a);

let init: len:int => f:(int => 'a) => list('a);

let append: list('a) => list('a) => list('a);

let rev_append: list('a) => list('a) => list('a);

let concat: list(list('a)) => list('a);

let flatten: list(list('a)) => list('a);

let equal: eq:('a => 'a => bool) => list('a) => list('a) => bool;

let compare: cmp:('a => 'a => int) => list('a) => list('a) => int;

let iter: f:('a => unit) => list('a) => unit;

let iteri: f:(int => 'a => unit) => list('a) => unit;

let map: f:('a => 'b) => list('a) => list('b);

let mapi: f:(int => 'a => 'b) => list('a) => list('b);

let rev_map: f:('a => 'b) => list('a) => list('b);

let filter_map: f:('a => option('b)) => list('a) => list('b);

let concat_map: f:('a => list('b)) => list('a) => list('b);

let fold_left_map: f:('acc => 'a => ('acc, 'b)) => init:'acc => list('a) => ('acc, list('b));

let fold_left: f:('acc => 'a => 'acc) => init:'acc => list('a) => 'acc;

let fold_right: f:('a => 'acc => 'acc) => list('a) => init:'acc => 'acc;

let iter2: f:('a => 'b => unit) => list('a) => list('b) => unit;

let map2: f:('a => 'b => 'c) => list('a) => list('b) => list('c);

let rev_map2: f:('a => 'b => 'c) => list('a) => list('b) => list('c);

let fold_left2: f:('acc => 'a => 'b => 'acc) => init:'acc => list('a) => list('b) => 'acc;

let fold_right2: f:('a => 'b => 'acc => 'acc) => list('a) => list('b) => init:'acc => 'acc;

let for_all: f:('a => bool) => list('a) => bool;

let exists: f:('a => bool) => list('a) => bool;

let for_all2: f:('a => 'b => bool) => list('a) => list('b) => bool;

let exists2: f:('a => 'b => bool) => list('a) => list('b) => bool;

let mem: 'a => set:list('a) => bool;

let memq: 'a => set:list('a) => bool;

let find: f:('a => bool) => list('a) => 'a;

let find_opt: f:('a => bool) => list('a) => option('a);

let find_index: f:('a => bool) => list('a) => option(int);

let find_map: f:('a => option('b)) => list('a) => option('b);

let find_mapi: f:(int => 'a => option('b)) => list('a) => option('b);

let filter: f:('a => bool) => list('a) => list('a);

let find_all: f:('a => bool) => list('a) => list('a);

let filteri: f:(int => 'a => bool) => list('a) => list('a);

let partition: f:('a => bool) => list('a) => (list('a), list('a));

let partition_map: f:('a => Stdlib.Either.t('b, 'c)) => list('a) => (list('b), list('c));

let assoc: 'a => list(('a, 'b)) => 'b;

let assoc_opt: 'a => list(('a, 'b)) => option('b);

let assq: 'a => list(('a, 'b)) => 'b;

let assq_opt: 'a => list(('a, 'b)) => option('b);

let mem_assoc: 'a => map:list(('a, 'b)) => bool;

let mem_assq: 'a => map:list(('a, 'b)) => bool;

let remove_assoc: 'a => list(('a, 'b)) => list(('a, 'b));

let remove_assq: 'a => list(('a, 'b)) => list(('a, 'b));

let split: list(('a, 'b)) => (list('a), list('b));

let combine: list('a) => list('b) => list(('a, 'b));

let sort: cmp:('a => 'a => int) => list('a) => list('a);

let stable_sort: cmp:('a => 'a => int) => list('a) => list('a);

let fast_sort: cmp:('a => 'a => int) => list('a) => list('a);

let sort_uniq: cmp:('a => 'a => int) => list('a) => list('a);

let merge: cmp:('a => 'a => int) => list('a) => list('a) => list('a);

let to_seq: list('a) => Stdlib.Seq.t('a);

let of_seq: Stdlib.Seq.t('a) => list('a);

let map_combine: list('a) => list('b) => ('a => 'c) => list(('c, 'b));

let map_combine_array: array('a) => list('b) => ('a => 'c) => list(('c, 'b));

let map_combine_array_append: array('a) => list('b) => list(('c, 'b)) => ('a => 'c) => list(('c, 'b));

let map_snd: list(('a, 'b)) => ('b => 'c) => list(('a, 'c));

let map_last: list('a) => (bool => 'a => 'b) => list('b);

map_last f xs will pass true to f for the last element, false otherwise. For empty list, it returns empty

let fold_left_with_offset: list('a) => 'acc => int => ('a => 'acc => int => 'acc) => 'acc;

let same_length: list('a) => list('b) => bool;

let split_at: list('a) => int => (list('a), list('a));

split_at n l will split l into two lists a,b, a will be of length n, otherwise, it will raise

let split_at_last: list('a) => (list('a), 'a);

split_at_last l is equivalent to split_at (List.length l - 1) l

let length_ge: list('a) => int => bool;

length xs = length ys + n 

input n should be positive TODO: input checking

let length_larger_than_n: list('a) => list('a) => int => bool;

let stable_group: list('a) => ('a => 'a => bool) => list(list('a));

stable_group eq lst Example: Input:

stable_group (=) [1;2;3;4;3]

Output:

[[1];[2];[4];[3;3]]

TODO: this is O(n^2) behavior which could be improved later

let rev_iter: list('a) => ('a => unit) => unit;

let for_all2_no_exn: list('a) => list('b) => ('a => 'b => bool) => bool;

for_all2_no_exn p xs ys returns true if all predicates satisfied, false otherwise or when the list lengths are not equal.

let split_map: list('a) => ('a => ('b, 'c)) => (list('b), list('c));

f is applied follow the list order