Module Parsing0

module Ast_helper: { ... };

Helpers to produce Parsetree fragments

module Ast_iterator: { ... };

Ast_iterator.iterator enables AST inspection using open recursion. A typical mapper would be based on Ast_iterator.default_iterator, a trivial iterator, and will fall back on it for handling the syntax it does not modify.

module Ast_mapper: { ... };

The interface of a -ppx rewriter

module Attr_helper: { ... };

Helpers for attributes

module Bs_clflags: { ... };

module Builtin_attributes: { ... };

Support for some of the builtin attributes

module Config: { ... };

System configuration

module Consistbl: { ... };

Consistency tables: for checking consistency of module CRCs

module Depend: { ... };

Module dependencies.

module Diffing: { ... };

Parametric diffing

module Diffing_with_keys: { ... };

When diffing lists where each element has a distinct key, we can refine the diffing patch by introducing two composite edit moves: swaps and moves.

module Docstrings: { ... };

Documentation comments

module Lazy_backtrack: { ... };

module Load_path: { ... };

Management of include directories.

module Location: { ... };

Source code locations (ranges of positions), used in parsetree.

module Misc: { ... };

Miscellaneous useful types and functions

module Parsetree: { ... };

Abstract syntax tree produced by parsing

module Printast: { ... };

Raw printer for Parsetree

module Warnings: { ... };

Warning definitions