Module Printf.Args

The Args module defines a heterogeneous list type, which can be used as the argument of the l*printf functions. For more documentation, see the similar module Format.Args.

type ('a, 'r) t =

| [] : ('r, 'r) t

| :: : 'a * ('b, 'r) t -> ('a -> 'b, 'r) t

val apply : 'a -> ('a, 'r) t -> 'r

val (@) : ('a, 'r1) t -> ('r1, 'r2) t -> ('a, 'r2) t