Module Stdlib.Unit

Unit values.

The unit type

type t = unit =

| ()

;

The unit type.

The constructor () is included here so that it has a path, but it is not intended to be used in user-defined data types.

let equal: t => t => bool;

equal u1 u2 is true.

let compare: t => t => int;

compare u1 u2 is 0.

let to_string: t => string;

to_string b is "()".