Module Ext_list

let map: list('a) => ('a => 'b) => list('b);

let map_combine: list('a) => list('b) => ('a => 'c) => list(('c, 'b));

let combine_array: array('a) => list('b) => ('a => 'c) => list(('c, 'b));

let combine_array_append: array('a) => list('b) => list(('c, 'b)) => ('a => 'c) => list(('c, 'b));

let has_string: list(string) => string => bool;

let mapi: list('a) => (int => 'a => 'b) => list('b);

let mapi_append: list('a) => (int => 'a => 'b) => list('b) => list('b);

let map_snd: list(('a, 'b)) => ('b => 'c) => list(('a, 'c));

let map_last: list('a) => (bool => 'a => 'b) => list('b);

map_last f xs will pass true to f for the last element, false otherwise. For empty list, it returns empty

let last: list('a) => 'a;

last l return the last element raise if the list is empty

let append: list('a) => list('a) => list('a);

let append_one: list('a) => 'a => list('a);

let map_append: list('b) => list('a) => ('b => 'a) => list('a);

let fold_right: list('a) => 'b => ('a => 'b => 'b) => 'b;

let fold_right2: list('a) => list('b) => 'c => ('a => 'b => 'c => 'c) => 'c;

let fold_right3: list('a) => list('b) => list('c) => 'd => ('a => 'b => 'c => 'd => 'd) => 'd;

let map2: list('a) => list('b) => ('a => 'b => 'c) => list('c);

let fold_left_with_offset: list('a) => 'acc => int => ('a => 'acc => int => 'acc) => 'acc;

let filter_map: list('a) => ('a => option('b)) => list('b);

let exclude_with_val: list('a) => ('a => bool) => option(list('a));

excludes p l return a tuple excluded,newl where exluded is true indicates that at least one element is removed,newl is the new list where all p x for x is false

let same_length: list('a) => list('b) => bool;

let init: int => (int => 'a) => list('a);

let split_at: list('a) => int => (list('a), list('a));

split_at n l will split l into two lists a,b, a will be of length n, otherwise, it will raise

let split_at_last: list('a) => (list('a), 'a);

split_at_last l It is equivalent to split_at (List.length l - 1) l

let filter_mapi: list('a) => ('a => int => option('b)) => list('b);

let length_ge: list('a) => int => bool;

length xs = length ys + n 

input n should be positive TODO: input checking

let length_larger_than_n: list('a) => list('a) => int => bool;

let rev_map_append: list('a) => list('b) => ('a => 'b) => list('b);

rev_map_append f l1 l2 map f l1 and reverse it to append l2 This weird semantics is due to it is the most efficient operation we can do

let flat_map: list('a) => ('a => list('b)) => list('b);

let stable_group: list('a) => ('a => 'a => bool) => list(list('a));

stable_group eq lst Example: Input:

stable_group (=) [1;2;3;4;3]

Output:

[[1];[2];[4];[3;3]]

TODO: this is O(n^2) behavior which could be improved later

let find_first: list('a) => ('a => bool) => option('a);

let find_first_exn: list('a) => ('a => bool) => 'a;

let find_first_not: list('a) => ('a => bool) => option('a);

find_first_not p lst if all elements in lst pass, return None otherwise return the first element e as Some e which fails the predicate

find_opt f l returns None if all return None, otherwise returns the first one.

let find_opt: list('a) => ('a => option('b)) => option('b);

let find_def: list('a) => ('a => option('b)) => 'b => 'b;

let rev_iter: list('a) => ('a => unit) => unit;

let iter: list('a) => ('a => unit) => unit;

let for_all: list('a) => ('a => bool) => bool;

let for_all_snd: list(('a, 'b)) => ('b => bool) => bool;

let for_all2_no_exn: list('a) => list('b) => ('a => 'b => bool) => bool;

for_all2_no_exn p xs ys return true if all satisfied, false otherwise or length not equal

let split_map: list('a) => ('a => ('b, 'c)) => (list('b), list('c));

f is applied follow the list order

let reduce_from_left: list('a) => ('a => 'a => 'a) => 'a;

fn is applied from left to right

let sort_via_array: list('a) => ('a => 'a => int) => list('a);

let sort_via_arrayf: list('a) => ('a => 'a => int) => ('a => 'b) => list('b);

let assoc_by_string: list((string, 'a)) => string => option('a) => 'a;

assoc_by_string default key lst if key is found in the list return that val, other unbox the default, otherwise assert false

let assoc_by_int: list((int, 'a)) => int => option('a) => 'a;

let nth_opt: list('a) => int => option('a);

let iter_snd: list(('a, 'b)) => ('b => unit) => unit;

let iter_fst: list(('a, 'b)) => ('a => unit) => unit;

let exists: list('a) => ('a => bool) => bool;

let exists_fst: list(('a, 'b)) => ('a => bool) => bool;

let exists_snd: list(('a, 'b)) => ('b => bool) => bool;

let concat_append: list(list('a)) => list('a) => list('a);

let fold_left2: list('a) => list('b) => 'c => ('a => 'b => 'c => 'c) => 'c;

let fold_left: list('a) => 'b => ('b => 'a => 'b) => 'b;

let singleton_exn: list('a) => 'a;

let filter: list('a) => ('a => bool) => list('a);