Module Melange_compiler_libs

module Annot : sig ... end

module Bs_clflags : sig ... end

module Btype : sig ... end

module Builtin_attributes : sig ... end

Support for the builtin attributes:

module Cmi_format : sig ... end

module Cmt_format : sig ... end

cmt and cmti files format.

module Config : sig ... end

System configuration

module Ctype : sig ... end

module Datarepr : sig ... end

module Env : sig ... end

module Envaux : sig ... end

module Errortrace : sig ... end

module Includeclass : sig ... end

module Includecore : sig ... end

module Includemod : sig ... end

module Includemod_errorprinter : sig ... end

module Lambda : sig ... end

module Location : sig ... end

Source code locations (ranges of positions), used in parsetree.

module Matching : sig ... end

module Misc : sig ... end

Miscellaneous useful types and functions

module Mtype : sig ... end

module Oprint : sig ... end

module Parmatch : sig ... end

Detection of partial matches and unused match cases.

module Patterns : sig ... end

module Persistent_env : sig ... end

module Predef : sig ... end

module Primitive : sig ... end

module Printast = Printast

module Printlambda : sig ... end

module Printpat : sig ... end

module Printtyp : sig ... end

Printing functions

module Out_type : sig ... end

Functions for representing type expressions and module types as outcometree (with as 'a aliases for cycles) and printing them. All functions below depends on global contexts that keep track of

module Printtyped : sig ... end

module Value_rec_check : sig ... end

module Value_rec_compiler : sig ... end

module Value_rec_types : sig ... end

Types related to the compilation of value let-recs (non-functional recursive definitions)

module Shape : sig ... end

Shapes are an abstract representation of modules' implementations which allow the tracking of definitions through functor applications and other module-level operations.

module Signature_group : sig ... end

Iterate on signature by syntactic group of items

module Stypes : sig ... end

module Subst : sig ... end

Substitutions

module Switch : sig ... end

module Tast_iterator : sig ... end

Allows the implementation of typed tree inspection using open recursion

module Tast_mapper : sig ... end

module Tmc : sig ... end

Tail-modulo-cons optimization.

module Translattribute : sig ... end

module Translclass : sig ... end

module Translcore : sig ... end

module Translmod : sig ... end

module Translobj : sig ... end

module Translprim : sig ... end

module Typeclass : sig ... end

module Typecore : sig ... end

module Typedecl : sig ... end

module Typedecl_immediacy : sig ... end

module Typedecl_properties : sig ... end

module Typedecl_separability : sig ... end

The OCaml runtime assumes for type-directed optimizations that all types are "separable". A type is "separable" if either all its inhabitants (the values of this type) are floating-point numbers, or none of them are.

module Typedecl_unboxed : sig ... end

module Typedecl_variance : sig ... end

module Typedtree : sig ... end

Abstract syntax tree after typing

module Typemod : sig ... end

Type-checking of the module language and typed ast hooks

module Typeopt : sig ... end

module Types : sig ... end

module Typetexp : sig ... end

module Unit_info : sig ... end

This module centralize the handling of compilation files and their metadata.

module Untypeast : sig ... end

module Warnings : sig ... end

Warning definitions