Module Melstd.Hash_set

Ideas are based on Hash, however, Hash.add does not really optimize and has a bad semantics for Hash_set, This module fixes the semantics of add. remove is not optimized since it is not used too much

module Make (H : Stdlib.Hashtbl.HashedType) : Hash_set_gen.S with type key = H.t

A naive t implementation on top of hashtbl, the value is unit